requirements.txt
## Module requirements
boto3
python-dateutil
websocket-client

## Development and testing requirements
coverage
pytest
pylint
build
twine
flake8
